+++ src/components/chart/chart.stories.tsx
@@ -1,15 +1,13 @@
 import { Meta, Story } from "@storybook/react/types-6-0";
-import { useState } from "react";
 import { useDarkMode } from "storybook-dark-mode";
 
 import { Interval } from "../../stories/api/vega-graphql";
 import { EmptyDataSource } from "../../stories/data-source/empty-data-source";
 import { JsonDataSource } from "../../stories/data-source/json-data-source";
-import { ThemeVariant } from "../../types";
 import { Chart, ChartProps } from "./chart";
 
 export default {
-  title: "Components/Chart",
+  title: "Charts/Chart",
   component: Chart,
 } as Meta;
 
@@ -60,4 +58,10 @@ export const SimpleMode = Template.bind({});
 SimpleMode.args = {
   ...Default.args,
   options: { studies: ["volume"], simple: true },
+};
+
+export const NoChange = Template.bind({});
+NoChange.args = {
+  dataSource: new JsonDataSource("", 5, "no-change-data.json", false),
+  interval: Interval.I1D,
 };