+++ src/math/array/disjoint.js
@@ -1,9 +1,10 @@
 export default function disjoint(values, other) {
-  const iterator = other[Symbol.iterator](), set = new Set();
+  const iterator = other[Symbol.iterator](),
+    set = new Set();
   for (const v of values) {
     if (set.has(v)) return false;
     let value, done;
-    while (({value, done} = iterator.next())) {
+    while (({ value, done } = iterator.next())) {
       if (done) break;
       if (Object.is(v, value)) return false;
       set.add(value);