+++ src/math/array/greatest.js
@@ -7,9 +7,9 @@ export default function greatest(values, compare = ascending) {
     let maxValue;
     for (const element of values) {
       const value = compare(element);
-      if (defined
-          ? ascending(value, maxValue) > 0
-          : ascending(value, value) === 0) {
+      if (
+        defined ? ascending(value, maxValue) > 0 : ascending(value, value) === 0
+      ) {
         max = element;
         maxValue = value;
         defined = true;
@@ -17,9 +17,7 @@ export default function greatest(values, compare = ascending) {
     }
   } else {
     for (const value of values) {
-      if (defined
-          ? compare(value, max) > 0
-          : compare(value, value) === 0) {
+      if (defined ? compare(value, max) > 0 : compare(value, value) === 0) {
         max = value;
         defined = true;
       }