+++ src/math/array/min.js
@@ -2,16 +2,20 @@ export default function min(values, valueof) {
   let min;
   if (valueof === undefined) {
     for (const value of values) {
-      if (value != null
-          && (min > value || (min === undefined && value >= value))) {
+      if (
+        value != null &&
+        (min > value || (min === undefined && value >= value))
+      ) {
         min = value;
       }
     }
   } else {
     let index = -1;
     for (let value of values) {
-      if ((value = valueof(value, ++index, values)) != null
-          && (min > value || (min === undefined && value >= value))) {
+      if (
+        (value = valueof(value, ++index, values)) != null &&
+        (min > value || (min === undefined && value >= value))
+      ) {
         min = value;
       }
     }