+++ src/math/array/minIndex.js
@@ -5,16 +5,20 @@ export default function minIndex(values, valueof) {
   if (valueof === undefined) {
     for (const value of values) {
       ++index;
-      if (value != null
-          && (min > value || (min === undefined && value >= value))) {
-        min = value, minIndex = index;
+      if (
+        value != null &&
+        (min > value || (min === undefined && value >= value))
+      ) {
+        (min = value), (minIndex = index);
       }
     }
   } else {
     for (let value of values) {
-      if ((value = valueof(value, ++index, values)) != null
-          && (min > value || (min === undefined && value >= value))) {
-        min = value, minIndex = index;
+      if (
+        (value = valueof(value, ++index, values)) != null &&
+        (min > value || (min === undefined && value >= value))
+      ) {
+        (min = value), (minIndex = index);
       }
     }
   }