+++ src/math/array/permute.js
@@ -1,3 +1,3 @@
-export default function(source, keys) {
-  return Array.from(keys, key => source[key]);
+export default function (source, keys) {
+  return Array.from(keys, (key) => source[key]);
 }