+++ src/math/array/sum.js
@@ -2,14 +2,14 @@ export default function sum(values, valueof) {
   let sum = 0;
   if (valueof === undefined) {
     for (let value of values) {
-      if (value = +value) {
+      if ((value = +value)) {
         sum += value;
       }
     }
   } else {
     let index = -1;
     for (let value of values) {
-      if (value = +valueof(value, ++index, values)) {
+      if ((value = +valueof(value, ++index, values))) {
         sum += value;
       }
     }