+++ src/math/array/variance.js
@@ -14,7 +14,10 @@ export default function variance(values, valueof) {
   } else {
     let index = -1;
     for (let value of values) {
-      if ((value = valueof(value, ++index, values)) != null && (value = +value) >= value) {
+      if (
+        (value = valueof(value, ++index, values)) != null &&
+        (value = +value) >= value
+      ) {
         delta = value - mean;
         mean += delta / ++count;
         sum += delta * (value - mean);