+++ src/types/theme.ts
@@ -1,3 +1,4 @@
 const darkTheme = "dark";
 const lightTheme = "light";
+
 export type ThemeVariant = typeof darkTheme | typeof lightTheme;